+++ contracts/ERC777TokensSender.sol
@@ -5,7 +5,6 @@
  * This code has not been reviewed.
  * Do not use or deploy this code before reviewing it personally first.
  */
-// solhint-disable-next-line compiler-fixed
 pragma solidity ^0.4.24;
 
 
@@ -14,7 +13,7 @@ interface ERC777TokensSender {
         address operator,
         address from,
         address to,
-        uint amount,
+        uint256 amount,
         bytes userData,
         bytes operatorData
     ) public;
